A car rental app built on multiple assumptions
Current assumptions:

Assuming account registration and other processes have been completed, you can log in directly.

Currently, the ignition status is used to determine if the order is in progress. Therefore, there may be two historical orders with a "in progress" status.

The map and historical orders use the same API to retrieve vehicle information. In a real project, these should be different APIs; here, we assume we are using two different APIs.

Continuing from the previous point, because we share the same API with the map, zooming the map will change the historical list. This wouldn't be an issue if we used two independently designed APIs.

Assuming that after successfully booking a vehicle, the map will only display the trajectory of the booked vehicle (using WebSocket), and the vehicle's color will change to yellow. Clicking on the vehicle will end the trip, and all vehicle information on the map will be retrieved again after the trip ends.

Because viewing historical orders simultaneously after a trip starts can cause some status inconsistencies (the demo assumes too many scenarios, potentially leading to very complex statuses), we assume that historical orders cannot be viewed after a successful booking until the trip ends.

Assuming the token has a longer expiration time, and the current token has a shorter expiration time, it will automatically log you back to the login page after expiration, allowing you to log in again.